refactor: keep proxy display helpers local
