+++ feeluown/app/app.py
@@ -29,8 +29,7 @@ from feeluown.plugin import plugins_mgr
 from feeluown.version import VersionManager
 from feeluown.task import TaskManager
 from feeluown.alert import AlertManager
-from feeluown.utils.utils import detect_proxy, format_proxies_for_display
-from feeluown.i18n import t
+from feeluown.utils.utils import detect_proxy
 
 from .mode import AppMode
 
@@ -128,13 +127,6 @@ class App:
         self.coll_mgr.scan()
         self.alert_mgr.initialize(self)
         self.proxies = detect_proxy()
-        if self.proxies:
-            logger.info(
-                t(
-                    "proxy-detected",
-                    proxy_info=format_proxies_for_display(self.proxies),
-                )
-            )
         self.player_pos_per300ms.initialize()
         self.player_pos_per300ms.changed.connect(self.live_lyric.on_position_changed)
         self.playlist.song_changed.connect(
