+++ tests/test_utils.py
@@ -1,11 +1,6 @@
 import pytest
 from copy import copy, deepcopy
-from feeluown.utils.utils import (
-    DedupList,
-    sanitize_proxy_url,
-    sanitize_proxies,
-    format_proxies_for_display,
-)
+from feeluown.utils.utils import DedupList
 
 
 class Obj:
@@ -125,35 +120,3 @@ def test_dedup_list():
     # clear
     dlist.clear()
     assert dlist == []
-
-
-def test_sanitize_proxy_url_removes_userinfo():
-    assert (
-        sanitize_proxy_url("http://user:pass@127.0.0.1:7890")
-        == "http://127.0.0.1:7890"
-    )
-
-
-def test_sanitize_proxies_keeps_mapping_shape():
-    proxies = {
-        "http": "http://user:pass@127.0.0.1:7890",
-        "https": "socks5://127.0.0.1:7891",
-    }
-
-    assert sanitize_proxies(proxies) == {
-        "http": "http://127.0.0.1:7890",
-        "https": "socks5://127.0.0.1:7891",
-    }
-
-
-def test_sanitize_proxy_url_keeps_ipv6_host_brackets():
-    assert (
-        sanitize_proxy_url("http://user:pass@[::1]:7890")
-        == "http://[::1]:7890"
-    )
-
-
-def test_format_proxies_for_display_uses_sanitized_values():
-    proxies = {"http": "http://user:pass@127.0.0.1:7890"}
-
-    assert format_proxies_for_display(proxies) == "http=http://127.0.0.1:7890"
